# mtms-static/contracts/openapi.yaml
openapi: 3.1.0
...
info:
  title: MTMS — Mahajan Ticket Management System
  version: 1.0.0
  description: |
    Getting change activities into production: an editable, attributable deliverable matrix.

    Permissions are enforced server-side on every request. The client holds the same
    permission keys only to disable controls and explain why.
servers:
  - url: /api/v1
...
security:
  - sessionCookie: []
paths:
...
  /config/columns:
    post:
      tags: [configuration]
      summary: Add a deliverable column
      description: |
        Appears at once on the matrix, in the readiness maths, on the module detail and in
        the audit, for every module. Requires `project.config`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties: { name: { type: string } }
      responses:
        '200': { $ref: '#/components/responses/Snapshot' }
  /config/columns/{key}:
    patch:
      tags: [configuration]
      summary: Change whether a column counts, which statuses it takes, or where it sits
      description: |
        Each field is applied only if present, so a screen can send whichever the user
        changed.

        **Changing `allowed` never rewrites cells.** A cell holding a status its column no
        longer permits keeps it, still counts toward readiness if its tone is done, and is
        counted into `off_vocabulary` so Configure can report it. Rewriting them would
        destroy the record of what was actually loaded, which is the one thing this
        application exists to protect.
      parameters: [{ name: key, in: path, required: true, schema: { type: string } }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              minProperties: 1
              properties:
                counts: { type: boolean }
                allowed: { type: array, items: { type: string }, minItems: 1 }
                move: { type: string, enum: [up, down] }
      responses:
        '200': { $ref: '#/components/responses/Snapshot' }
    delete:
      tags: [configuration]
      summary: Remove a column
      description: Its cells go with it, so a reused key cannot resurrect old values.
      parameters: [{ name: key, in: path, required: true, schema: { type: string } }]
      responses:
        '200': { $ref: '#/components/responses/Snapshot' }
  /config/lists:
    post:
      tags: [configuration]
      summary: Edit node types, pipeline stages, owners or link types
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [list, action, value]
              properties:
                list: { type: string, enum: [node_types, stages, owners, link_types] }
                action: { type: string, enum: [add, remove] }
                value: { type: string }
      responses:
        '200': { $ref: '#/components/responses/Snapshot' }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: tracker_at
      description: Short-lived access token. The tenant claim inside it is the only source of tenancy.
    ingestToken:
      type: http
      scheme: bearer
      description: '`DRIFT_INGEST_TOKEN`, for agents that have no session.'
...
  responses:
    Snapshot:
      description: The whole project, as it now stands
      content:
        application/json:
          schema:
            type: object
            properties:
              data: { $ref: '#/components/schemas/Snapshot' }
              meta: { type: object, additionalProperties: true }
...
  schemas:
...
    PermissionKey:
      type: string
      description: The single vocabulary of authorisation, enforced server-side on every request.
      enum:
        - project.view
        - project.create
        - project.members.manage
        - project.config
        - module.create
        - module.edit
        - module.clone
        - deliverable.update
        - prod.confirm
        - fni.date
        - fni.signoff
        - defect.create
        - defect.transition
        - defect.assign
        - admin.users.manage
        - admin.roles.manage
        - admin.audit.view
    StatusKey:
      type: string
      description: |
        The shared status vocabulary. Each column is configured with the subset it may use.

        `""` is a distinct value meaning **nothing recorded** — it is not a status. It has
        its own tone, renders conspicuously, and counts as a gap on the dashboard. Never
        coerce it to "not done".
      enum: ['', notcreated, created, notloaded, lab, prod, loaded, pending, completed, notraised, raised]
...
    CellView:
      type: object
      required: [column_key, status, rolled_up]
      properties:
        column_key: { type: string }
        status: { $ref: '#/components/schemas/StatusKey' }
        rolled_up:
          type: boolean
          description: Derived from subactivities. Must not be edited directly; the API refuses it.
        subactivity_count: { type: integer }
        changed_by: { type: [string, 'null'] }
        changed_at: { type: [string, 'null'], format: date-time }
    ModuleView:
      type: object
      description: A module is a node type plus an activity, and that pair is its identity.
      properties:
        id: { type: string, format: uuid }
        node_type: { type: string }
        name: { type: string }
        owner: { type: [string, 'null'] }
        fni_target_date: { type: [string, 'null'], format: date }
        closed: { type: boolean }
        closed_by: { type: [string, 'null'] }
        readiness:
          type: integer
          minimum: 0
          maximum: 100
          description: Done cells ÷ counted columns. Derived, never stored.
        stage_index: { type: integer, description: Bucketed from readiness across the configured stages. }
        missing: { type: array, items: { type: string } }
        blank_count: { type: integer }
        cells: { type: array, items: { $ref: '#/components/schemas/CellView' } }
        subactivities: { type: array, items: { type: object } }
        links: { type: array, items: { type: object } }
        last_run: { type: [object, 'null'] }
    Snapshot:
      type: object
      description: |
        Everything the ten screens read. Every mutation returns a fresh one, so the client
        never has to merge a partial update.
      properties:
        me:
          type: object
          properties:
            user_id: { type: string, format: uuid }
            display_name: { type: string }
            email: { type: string }
            role_names: { type: array, items: { type: string } }
            permissions: { type: array, items: { $ref: '#/components/schemas/PermissionKey' } }
        org: { type: object }
        project: { type: object }
        projects: { type: array, items: { type: object } }
        config: { type: object }
        modules: { type: array, items: { $ref: '#/components/schemas/ModuleView' } }
        audit: { type: array, items: { type: object } }
        defects: { type: array, items: { type: object } }
        library: { type: array, items: { type: object } }
        roles: { type: array, items: { type: object } }
        users: { type: array, items: { type: object } }
        members: { type: array, items: { type: object } }
        invitations: { type: array, items: { type: object } }
        drift:
          type: object
          properties:
            rows: { type: array, items: { type: object } }
            warnings: { type: array, items: { type: object } }
            gate: { type: object }
            reports: { type: array, items: { type: object } }
            promotions: { type: array, items: { type: object } }